move subcomponents into subfolders
